refactor: users module
