+++ src/modules/users/users.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common'
-import { UsersResolver } from './users.resolver'
-import { UsersService } from './users.service'
+import { UsersResolver } from './resolvers/users.resolver'
+import { UsersService } from './services/users.service'
 
 @Module({
   providers: [UsersResolver, UsersService],
